feat(fs-simd): add gated SME2 exploratory capsule
Add the first SME2 streaming-mode prototype as an explicitly frontier-gated
unsafe capsule:
- introduce the `frontier-sme2` feature on `fs-simd` and keep the module gated
  to aarch64, non-Miri builds
- add a runtime capability probe that requires OS-reported SME2 support and the
  fixed 64-byte streaming vector length expected by the 16x16 f32 prototype
- add a scalar twin with identical panel layout and accumulation order so same
  hardware equivalence can be measured before any future promotion
- register the capsule in `unsafe-capsules.json` and add the dedicated SAFETY.md
  describing streaming-mode containment, register discipline, bounds, capability
  gating, compensating checks, and residual SME2 maturity risk
- remove a redundant cast from the topopt marquee seeded-design fixture while
  preserving the same floating-point expression

Validation:
- rustfmt --edition 2024 --check crates/fs-simd/src/sme2/mod.rs
- cargo check -p fs-simd --features frontier-sme2
- cargo run -p xtask -- check-unsafe
- cargo test -p fs-topopt --features cutfem-marquee --no-run
- git diff --check -- crates/fs-simd/Cargo.toml crates/fs-simd/src/lib.rs crates/fs-simd/src/sme2/mod.rs crates/fs-simd/src/sme2/SAFETY.md unsafe-capsules.json crates/fs-topopt/tests/marquee.rs
- ubs crates/fs-simd/src/sme2/mod.rs crates/fs-topopt/tests/marquee.rs reported 0 critical issues

+++ crates/fs-topopt/tests/marquee.rs
@@ -30,7 +30,7 @@ fn seeded_design() -> DensityDesign {
         for i in 0..n {
             let (x, y) = (i as f64 / scale, j as f64 / scale);
             for &(cx, cy) in &[(0.3, 0.5), (0.7, 0.5)] {
-                if ((x - cx) as f64).powi(2) + ((y - cy) as f64).powi(2) < 0.012 {
+                if (x - cx).powi(2) + (y - cy).powi(2) < 0.012 {
                     d.rho[j * n + i] = 0.1;
                 }
             }
